Omnitree Update
The Equate<T> was only necessary for the Remove(T x) method, so it was moved to an optional parameter on that method rather than a field on the class.

Also, Type Classes are not a thing in C# yet, so I did some extension method hacking to route the remove methods into potentially shared functions on the Omnitree static class.